requirements.txt
pandas
plotly_express
streamlit
nbformat>=4.2.0